requirements.txt
annotated-types==0.7.0
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docstring_parser==0.16
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.4
grpcio-status==1.62.3
idna==3.7
jupyter_client==8.6.2
jupyter_core==5.7.2
kfp==2.8.0
kfp-pipeline-spec==0.3.0
kfp-server-api==2.0.5
kubernetes==26.1.0
numpy==2.0.1
oauthlib==3.2.2
packaging==24.1
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
pywin32==306
PyYAML==6.0.1
pyzmq==26.1.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9
shapely==2.0.5
six==1.16.0
tabulate==0.9.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==1.26.19
websocket-client==1.8.0
